pyproject.toml
[project]
name = "openapi-mcp-server"
version = "0.1.0"
description = "MCP server for OpenAPI specifications"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "mcp[cli]>=1.9.0",
    "PyYAML==6.0.1",
    "httpx>=0.24.0",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true 